requirements.txt
# 🌊 A-GENTEE v6.0 — Cloud Backend Dependencies
# For Railway deployment

# Web framework
fastapi==0.115.*
uvicorn[standard]==0.34.*

# AI Engines
anthropic>=0.40.0
google-genai>=1.0.0
openai>=1.50.0

# HTTP client (for Supabase REST + ElevenLabs)
httpx>=0.27.0

# Voice
edge-tts>=6.1.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
